stories-inbox: move sync-trigger creation to a TriggersModule toolbar button
Replace the decorator's reactive auto-install with a manual "Create trigger"
toolbar action in TriggersModule (creates the GoogleMailSync trigger + relation
for the mailbox's binding); simplify the MailboxSync ClientPlugin initializer.
